+++ src/vqueue/src/main.rs
@@ -33,11 +33,8 @@ async fn main() -> anyhow::Result<()> {
     }
 
     if let Some(command) = args.command {
-        let config = args.config.as_ref().map_or_else(
-            || Ok(Config::default()),
-            |path| Config::from_vsl_file(path).context("Cannot parse the configuration"),
-        )?;
-
+        let config =
+            Config::from_vsl_file(args.config).context("Cannot parse the configuration")?;
         let config = std::sync::Arc::new(config);
         let manager = vqueue::fs::QueueManager::init(config, vec![])?;